// src/schema.ts
export interface IContentType {
    Name: string;
    StringId: string;
    ParentStringId?: string;
    Description?: string;
    Group?: string;
    FieldRefs?: IContentTypeFieldRef[];
}

export interface IContentTypeFieldRef {
    FieldInternalName: string;
}